+++ react-accordion-component/src/accordion.jsx
@@ -1,37 +1,70 @@
-// import React from 'react';
+import React from 'react';
 
-// function TestComp(props) {
+class Numbers extends React.Component {
 
-//   return (
-//     <p>{props}</p>
-//   );
-// }
+  constructor(props) {
+    super(props);
+    this.clickHandle = this.clickHandle.bind(this);
+  }
 
-// function click() {
-//   return true;
-// }
+  clickHandle(event) {
 
-// function Accordian(props) {
-//   const subBoxes = (
-//     <div>
-//       {props.subjects.map(lang =>
-//         <div key={lang.lang}>
-//           {lang.lang}
-//         </div>
-//       )}
-//     </div>
+    const targetname = event.target.getAttribute('id');
+    const targetHTML = document.querySelector('.description.' + targetname);
+    const otherHTML = document.querySelectorAll('.description');
 
-//   );
+    if (targetHTML.classList.contains('hidden')) {
 
-//   return (
-//     <div >
-//       {subBoxes}
-//       {/* <TestComp /> */}
-//     </div>
-//   );
-// }
-// export default Accordian;
+      targetHTML.classList.remove('hidden');
+      for (let i = 0; i < otherHTML.length; i++) {
+        if (!otherHTML[i].classList.contains(targetname)) {
+          if (!otherHTML[i].classList.contains('hidden')) {
+            otherHTML[i].classList.add('hidden');
+          }
 
-// class
+        }
+      }
+    } else {
+      targetHTML.classList.add('hidden');
+    }
 
-// export default Accordian;
+  }
+
+  render() {
+    const lang = this.props.lang;
+    return (
+      <div>
+
+        {lang.map(langIt =>
+
+          <div className="name-div hover" id={langIt.lang} key={langIt.lang} onClick={this.clickHandle}>
+
+            {langIt.lang}
+            <div className={'description ' + langIt.lang + ' hidden'}>
+
+              {langIt.description}
+            </div>
+          </div>
+
+        )}
+
+      </div>
+    );
+  }
+}
+
+class Accordion extends React.Component {
+
+  render() {
+    const input = this.props.input;
+    return (
+      <div>
+
+        <Numbers lang={input} />
+
+      </div>
+    );
+  }
+}
+
+export default Accordion;
